Game Playable (Movemend and Attack added)

+++ src/business/Vector2.java
@@ -1,10 +1,10 @@
 package business;
 
 public class Vector2 {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Vector2(int x, int y){
+    public Vector2(double x, double y){
         this.x = x;
         this.y = y;
     }
